+++ web/eslint.config.mjs
@@ -7,4 +7,9 @@ const compat = new FlatCompat({
   baseDirectory: __dirname,
 });
 
-export default [...compat.extends("next/core-web-vitals")];
+const config = [
+  { ignores: [".next/**"] },
+  ...compat.extends("next/core-web-vitals")
+];
+
+export default config;